+++ spec/dummy/script/register_client_secret_post.rb
@@ -12,8 +12,8 @@ require "json"
 require "uri"
 
 # Ensure dynamic registration is enabled
-unless TokenAuthority.config.rfc_7591_enabled
-  TokenAuthority.config.rfc_7591_enabled = true
+unless TokenAuthority.config.dcr_enabled
+  TokenAuthority.config.dcr_enabled = true
   puts "Enabled dynamic client registration for this session.\n\n"
 end