+++ spec/models/token_authority/protected_resource_metadata_spec.rb
@@ -5,13 +5,13 @@ require "rails_helper"
 RSpec.describe TokenAuthority::ProtectedResourceMetadata, type: :model do
   let!(:original_config) do
     {
-      rfc_9068_issuer_url: TokenAuthority.config.rfc_9068_issuer_url,
+      token_issuer_url: TokenAuthority.config.token_issuer_url,
       resources: TokenAuthority.config.resources
     }
   end
 
   before do
-    TokenAuthority.config.rfc_9068_issuer_url = "https://example.com/"
+    TokenAuthority.config.token_issuer_url = "https://example.com/"
     TokenAuthority.config.resources = {}
   end